+++ src/commands/autocorrectAll.ts
@@ -1,10 +1,9 @@
 import * as vscode from 'vscode';
 import { executeCommandsInTerminal } from './utils';
 import { SorbetExtensionContext } from '../sorbetExtensionContext';
-import { DEFAULT_SORBET_TYPECHECK } from '../configuration/configuration';
 
-export async function autocorrectAll(_context: SorbetExtensionContext, code: string | number, contextUri: vscode.Uri) {
-  const sorbetCommand = DEFAULT_SORBET_TYPECHECK.join(' ');
+export async function autocorrectAll({ configuration }: SorbetExtensionContext, code: string | number, contextUri: vscode.Uri) {
+  const sorbetCommand = configuration.sorbetTypecheckCommand.join(' ').trim();
   await executeCommandsInTerminal(
     {
       commands: [`${sorbetCommand} --autocorrect --isolate-error-code=${code}`],